refactor: improve policycache package
