+++ pkg/policycache/cache_test.go
@@ -49,7 +49,7 @@ func Test_All(t *testing.T) {
 	pCache := newPolicyCache(log.Log, dummyLister{}, dummyNsLister{})
 	policy := newPolicy(t)
 	//add
-	pCache.Add(policy)
+	pCache.add(policy)
 	for _, rule := range autogen.ComputeRules(policy) {
 		for _, kind := range rule.MatchResources.Kinds {
 
@@ -71,7 +71,7 @@ func Test_All(t *testing.T) {
 	}
 
 	// remove
-	pCache.Remove(policy)
+	pCache.remove(policy)
 	kind := "pod"
 	validateEnforce := pCache.get(ValidateEnforce, kind, "")
 	assert.Assert(t, len(validateEnforce) == 0)
@@ -80,9 +80,9 @@ func Test_All(t *testing.T) {
 func Test_Add_Duplicate_Policy(t *testing.T) {
 	pCache := newPolicyCache(log.Log, dummyLister{}, dummyNsLister{})
 	policy := newPolicy(t)
-	pCache.Add(policy)
-	pCache.Add(policy)
-	pCache.Add(policy)
+	pCache.add(policy)
+	pCache.add(policy)
+	pCache.add(policy)
 	for _, rule := range autogen.ComputeRules(policy) {
 		for _, kind := range rule.MatchResources.Kinds {
 
@@ -106,12 +106,12 @@ func Test_Add_Duplicate_Policy(t *testing.T) {
 func Test_Add_Validate_Audit(t *testing.T) {
 	pCache := newPolicyCache(log.Log, dummyLister{}, dummyNsLister{})
 	policy := newPolicy(t)
-	pCache.Add(policy)
-	pCache.Add(policy)
+	pCache.add(policy)
+	pCache.add(policy)
 
 	policy.Spec.ValidationFailureAction = "audit"
-	pCache.Add(policy)
-	pCache.Add(policy)
+	pCache.add(policy)
+	pCache.add(policy)
 	for _, rule := range autogen.ComputeRules(policy) {
 		for _, kind := range rule.MatchResources.Kinds {
 
@@ -132,7 +132,7 @@ func Test_Add_Remove(t *testing.T) {
 	pCache := newPolicyCache(log.Log, dummyLister{}, dummyNsLister{})
 	policy := newPolicy(t)
 	kind := "Pod"
-	pCache.Add(policy)
+	pCache.add(policy)
 
 	validateEnforce := pCache.get(ValidateEnforce, kind, "")
 	if len(validateEnforce) != 1 {
@@ -149,7 +149,7 @@ func Test_Add_Remove(t *testing.T) {
 		t.Errorf("expected 1 generate policy, found %v", len(generate))
 	}
 
-	pCache.Remove(policy)
+	pCache.remove(policy)
 	deletedValidateEnforce := pCache.get(ValidateEnforce, kind, "")
 	if len(deletedValidateEnforce) != 0 {
 		t.Errorf("expected 0 validate enforce policy, found %v", len(deletedValidateEnforce))
@@ -160,7 +160,7 @@ func Test_Add_Remove_Any(t *testing.T) {
 	pCache := newPolicyCache(log.Log, dummyLister{}, dummyNsLister{})
 	policy := newAnyPolicy(t)
 	kind := "Pod"
-	pCache.Add(policy)
+	pCache.add(policy)
 
 	validateEnforce := pCache.get(ValidateEnforce, kind, "")
 	if len(validateEnforce) != 1 {
@@ -177,7 +177,7 @@ func Test_Add_Remove_Any(t *testing.T) {
 		t.Errorf("expected 1 generate policy, found %v", len(generate))
 	}
 
-	pCache.Remove(policy)
+	pCache.remove(policy)
 	deletedValidateEnforce := pCache.get(ValidateEnforce, kind, "")
 	if len(deletedValidateEnforce) != 0 {
 		t.Errorf("expected 0 validate enforce policy, found %v", len(deletedValidateEnforce))
@@ -188,7 +188,7 @@ func Test_Remove_From_Empty_Cache(t *testing.T) {
 	pCache := newPolicyCache(log.Log, nil, nil)
 	policy := newPolicy(t)
 
-	pCache.Remove(policy)
+	pCache.remove(policy)
 }
 
 func newPolicy(t *testing.T) *kyverno.ClusterPolicy {
@@ -929,7 +929,7 @@ func Test_Ns_All(t *testing.T) {
 	pCache := newPolicyCache(log.Log, dummyLister{}, dummyNsLister{})
 	policy := newNsPolicy(t)
 	//add
-	pCache.Add(policy)
+	pCache.add(policy)
 	nspace := policy.GetNamespace()
 	for _, rule := range autogen.ComputeRules(policy) {
 		for _, kind := range rule.MatchResources.Kinds {
@@ -951,7 +951,7 @@ func Test_Ns_All(t *testing.T) {
 		}
 	}
 	// remove
-	pCache.Remove(policy)
+	pCache.remove(policy)
 	kind := "pod"
 	validateEnforce := pCache.get(ValidateEnforce, kind, nspace)
 	assert.Assert(t, len(validateEnforce) == 0)
@@ -960,9 +960,9 @@ func Test_Ns_All(t *testing.T) {
 func Test_Ns_Add_Duplicate_Policy(t *testing.T) {
 	pCache := newPolicyCache(log.Log, dummyLister{}, dummyNsLister{})
 	policy := newNsPolicy(t)
-	pCache.Add(policy)
-	pCache.Add(policy)
-	pCache.Add(policy)
+	pCache.add(policy)
+	pCache.add(policy)
+	pCache.add(policy)
 	nspace := policy.GetNamespace()
 	for _, rule := range autogen.ComputeRules(policy) {
 		for _, kind := range rule.MatchResources.Kinds {
@@ -987,12 +987,12 @@ func Test_Ns_Add_Duplicate_Policy(t *testing.T) {
 func Test_Ns_Add_Validate_Audit(t *testing.T) {
 	pCache := newPolicyCache(log.Log, dummyLister{}, dummyNsLister{})
 	policy := newNsPolicy(t)
-	pCache.Add(policy)
-	pCache.Add(policy)
+	pCache.add(policy)
+	pCache.add(policy)
 	nspace := policy.GetNamespace()
 	policy.Spec.ValidationFailureAction = "audit"
-	pCache.Add(policy)
-	pCache.Add(policy)
+	pCache.add(policy)
+	pCache.add(policy)
 	for _, rule := range autogen.ComputeRules(policy) {
 		for _, kind := range rule.MatchResources.Kinds {
 
@@ -1014,13 +1014,13 @@ func Test_Ns_Add_Remove(t *testing.T) {
 	policy := newNsPolicy(t)
 	nspace := policy.GetNamespace()
 	kind := "Pod"
-	pCache.Add(policy)
+	pCache.add(policy)
 	validateEnforce := pCache.get(ValidateEnforce, kind, nspace)
 	if len(validateEnforce) != 1 {
 		t.Errorf("expected 1 validate enforce policy, found %v", len(validateEnforce))
 	}
 
-	pCache.Remove(policy)
+	pCache.remove(policy)
 	deletedValidateEnforce := pCache.get(ValidateEnforce, kind, nspace)
 	if len(deletedValidateEnforce) != 0 {
 		t.Errorf("expected 0 validate enforce policy, found %v", len(deletedValidateEnforce))
@@ -1031,7 +1031,7 @@ func Test_GVk_Cache(t *testing.T) {
 	pCache := newPolicyCache(log.Log, dummyLister{}, dummyNsLister{})
 	policy := newGVKPolicy(t)
 	//add
-	pCache.Add(policy)
+	pCache.add(policy)
 	for _, rule := range autogen.ComputeRules(policy) {
 		for _, kind := range rule.MatchResources.Kinds {
 
@@ -1047,13 +1047,13 @@ func Test_GVK_Add_Remove(t *testing.T) {
 	pCache := newPolicyCache(log.Log, dummyLister{}, dummyNsLister{})
 	policy := newGVKPolicy(t)
 	kind := "ClusterRole"
-	pCache.Add(policy)
+	pCache.add(policy)
 	generate := pCache.get(Generate, kind, "")
 	if len(generate) != 1 {
 		t.Errorf("expected 1 generate policy, found %v", len(generate))
 	}
 
-	pCache.Remove(policy)
+	pCache.remove(policy)
 	deletedGenerate := pCache.get(Generate, kind, "")
 	if len(deletedGenerate) != 0 {
 		t.Errorf("expected 0 generate policy, found %v", len(deletedGenerate))
@@ -1065,7 +1065,7 @@ func Test_Add_Validate_Enforce(t *testing.T) {
 	policy := newUserTestPolicy(t)
 	nspace := policy.GetNamespace()
 	//add
-	pCache.Add(policy)
+	pCache.add(policy)
 	for _, rule := range autogen.ComputeRules(policy) {
 		for _, kind := range rule.MatchResources.Kinds {
 			validateEnforce := pCache.get(ValidateEnforce, kind, nspace)
@@ -1081,13 +1081,13 @@ func Test_Ns_Add_Remove_User(t *testing.T) {
 	policy := newUserTestPolicy(t)
 	nspace := policy.GetNamespace()
 	kind := "Deployment"
-	pCache.Add(policy)
+	pCache.add(policy)
 	validateEnforce := pCache.get(ValidateEnforce, kind, nspace)
 	if len(validateEnforce) != 1 {
 		t.Errorf("expected 1 validate enforce policy, found %v", len(validateEnforce))
 	}
 
-	pCache.Remove(policy)
+	pCache.remove(policy)
 	deletedValidateEnforce := pCache.get(ValidateEnforce, kind, nspace)
 	if len(deletedValidateEnforce) != 0 {
 		t.Errorf("expected 0 validate enforce policy, found %v", len(deletedValidateEnforce))
@@ -1098,9 +1098,9 @@ func Test_Mutate_Policy(t *testing.T) {
 	pCache := newPolicyCache(log.Log, dummyLister{}, dummyNsLister{})
 	policy := newMutatePolicy(t)
 	//add
-	pCache.Add(policy)
-	pCache.Add(policy)
-	pCache.Add(policy)
+	pCache.add(policy)
+	pCache.add(policy)
+	pCache.add(policy)
 	for _, rule := range autogen.ComputeRules(policy) {
 		for _, kind := range rule.MatchResources.Kinds {
 
@@ -1117,7 +1117,7 @@ func Test_Generate_Policy(t *testing.T) {
 	pCache := newPolicyCache(log.Log, dummyLister{}, dummyNsLister{})
 	policy := newgenratePolicy(t)
 	//add
-	pCache.Add(policy)
+	pCache.add(policy)
 	for _, rule := range autogen.ComputeRules(policy) {
 		for _, kind := range rule.MatchResources.Kinds {
 
@@ -1135,10 +1135,10 @@ func Test_NsMutate_Policy(t *testing.T) {
 	policy := newMutatePolicy(t)
 	nspolicy := newNsMutatePolicy(t)
 	//add
-	pCache.Add(policy)
-	pCache.Add(nspolicy)
-	pCache.Add(policy)
-	pCache.Add(nspolicy)
+	pCache.add(policy)
+	pCache.add(nspolicy)
+	pCache.add(policy)
+	pCache.add(nspolicy)
 
 	nspace := policy.GetNamespace()
 	// get
@@ -1159,8 +1159,8 @@ func Test_Validate_Enforce_Policy(t *testing.T) {
 	pCache := newPolicyCache(log.Log, dummyLister{}, dummyNsLister{})
 	policy1 := newValidateAuditPolicy(t)
 	policy2 := newValidateEnforcePolicy(t)
-	pCache.Add(policy1)
-	pCache.Add(policy2)
+	pCache.add(policy1)
+	pCache.add(policy2)
 
 	validateEnforce := pCache.get(ValidateEnforce, "Pod", "")
 	if len(validateEnforce) != 2 {
@@ -1172,8 +1172,8 @@ func Test_Validate_Enforce_Policy(t *testing.T) {
 		t.Errorf("adding: expected 0 validate audit policy, found %v", len(validateAudit))
 	}
 
-	pCache.Remove(policy1)
-	pCache.Remove(policy2)
+	pCache.remove(policy1)
+	pCache.remove(policy2)
 
 	validateEnforce = pCache.get(ValidateEnforce, "Pod", "")
 	if len(validateEnforce) != 0 {
